PYI-788: Rename everything and change endpoints
This renames a lot of stuff to be (hopefully) more intuitively named.

It also changes the endpoints on the API gateway for passport-back -
this will need reflecting in the frontend and in the config.

Of particular note is the change of the endpoint the credentials are
issued on. This has change to a POST to /credentials/issue. This is to
fit in with the vc http api spec, as per ticket PYI-788.
