--- lambdas/dcscredential/src/main/java/uk/gov/di/ipv/cri/passport/dcscredential/DcsCredentialHandler.java
@@ -1,85 +1,0 @@
-package uk.gov.di.ipv.cri.passport.dcscredential;
-
-import com.amazonaws.services.lambda.runtime.Context;
-import com.amazonaws.services.lambda.runtime.RequestHandler;
-import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
-import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
-import com.nimbusds.oauth2.sdk.OAuth2Error;
-import com.nimbusds.oauth2.sdk.ParseException;
-import com.nimbusds.oauth2.sdk.token.AccessToken;
-import com.nimbusds.oauth2.sdk.util.StringUtils;
-import org.apache.http.HttpStatus;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import uk.gov.di.ipv.cri.passport.dcscredential.domain.PassportCredentialIssuerResponse;
-import uk.gov.di.ipv.cri.passport.library.annotations.ExcludeFromGeneratedCoverageReport;
-import uk.gov.di.ipv.cri.passport.library.helpers.ApiGatewayResponseGenerator;
-import uk.gov.di.ipv.cri.passport.library.helpers.RequestHelper;
-import uk.gov.di.ipv.cri.passport.library.persistence.item.PassportCheckDao;
-import uk.gov.di.ipv.cri.passport.library.service.AccessTokenService;
-import uk.gov.di.ipv.cri.passport.library.service.ConfigurationService;
-import uk.gov.di.ipv.cri.passport.library.service.DcsCredentialService;
-
-public class DcsCredentialHandler
-        implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(DcsCredentialHandler.class);
-    private static final String AUTHORIZATION_HEADER_KEY = "Authorization";
-
-    private final DcsCredentialService dcsCredentialService;
-    private final AccessTokenService accessTokenService;
-    private final ConfigurationService configurationService;
-
-    public DcsCredentialHandler(
-            DcsCredentialService dcsCredentialService,
-            AccessTokenService accessTokenService,
-            ConfigurationService configurationService) {
-        this.configurationService = configurationService;
-        this.dcsCredentialService = dcsCredentialService;
-        this.accessTokenService = accessTokenService;
-    }
-
-    @ExcludeFromGeneratedCoverageReport
-    public DcsCredentialHandler() {
-        this.configurationService = new ConfigurationService();
-        this.dcsCredentialService = new DcsCredentialService(configurationService);
-        this.accessTokenService = new AccessTokenService(configurationService);
-    }
-
-    @Override
-    public APIGatewayProxyResponseEvent handleRequest(
-            APIGatewayProxyRequestEvent input, Context context) {
-        try {
-            String accessTokenString =
-                    RequestHelper.getHeaderByKey(input.getHeaders(), AUTHORIZATION_HEADER_KEY);
-
-            // Performs validation on header value and throws a ParseException if invalid
-            AccessToken.parse(accessTokenString);
-
-            String resourceId = accessTokenService.getResourceIdByAccessToken(accessTokenString);
-
-            if (StringUtils.isBlank(resourceId)) {
-                LOGGER.error(
-                        "User credential could not be retrieved. The supplied access token was not found in the database.");
-                return ApiGatewayResponseGenerator.proxyJsonResponse(
-                        OAuth2Error.ACCESS_DENIED.getHTTPStatusCode(),
-                        OAuth2Error.ACCESS_DENIED
-                                .appendDescription(
-                                        " - The supplied access token was not found in the database")
-                                .toJSONObject());
-            }
-
-            PassportCheckDao credential = dcsCredentialService.getDcsCredential(resourceId);
-
-            PassportCredentialIssuerResponse passportCredentialIssuerResponse =
-                    PassportCredentialIssuerResponse.fromPassportCheckDao(credential);
-
-            return ApiGatewayResponseGenerator.proxyJsonResponse(
-                    HttpStatus.SC_OK, passportCredentialIssuerResponse);
-        } catch (ParseException e) {
-            LOGGER.error("Failed to parse access token");
-            return ApiGatewayResponseGenerator.proxyJsonResponse(
-                    e.getErrorObject().getHTTPStatusCode(), e.getErrorObject().toJSONObject());
-        }
-    }
-}
