--- lambdas/passport/src/test/java/uk/gov/di/ipv/cri/passport/passport/OAuth2RequestParams.java
@@ -1,8 +1,0 @@
-package uk.gov.di.ipv.cri.passport.passport;
-
-public class OAuth2RequestParams {
-    public static final String REDIRECT_URI = "redirect_uri";
-    public static final String CLIENT_ID = "client_id";
-    public static final String RESPONSE_TYPE = "response_type";
-    public static final String SCOPE = "scope";
-}
