+++ lambdas/sharedattributes/build.gradle
